+++ tilings/algorithms/__init__.py
@@ -6,6 +6,7 @@ from .factor import (Factor, FactorWithInterleaving,
                      FactorWithMonotoneInterleaving)
 from .fusion import ComponentFusion, Fusion
 from .minimal_gridded_perms import MinimalGriddedPerms
+from .gridded_perm_generation import GriddedPermsOnTiling
 from .obstruction_inferral import (AllObstructionInferral, EmptyCellInferral,
                                    SubobstructionInferral)
 from .obstruction_transitivity import ObstructionTransitivity
@@ -36,6 +37,7 @@ __all__ = [
     'CellInsertion',
     'CrossingInsertion',
     'FactorInsertion',
+    'GriddedPermsOnTiling',
     'RequirementCorroboration',
     'RequirementExtension',
     'RequirementPlacement',